+++ skywalking/meter/meter.py
@@ -14,9 +14,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-
 from abc import ABC, abstractmethod
 from enum import Enum
+from typing import Optional
+
 from skywalking.protocol.language_agent.Meter_pb2 import Label
 import skywalking.meter as meter
 
@@ -82,9 +83,10 @@ class BaseMeter(ABC):
     meter_service = None
 
     def __init__(self, name: str, tags=None):
-        if BaseMeter.meter_service is None:
-            BaseMeter.meter_service = meter._meter_service
-
+        # Should always override to use the correct meter service.
+        # Otherwise, forked process will inherit the original
+        # meter_service in parent. We want a new one in child.
+        BaseMeter.meter_service = meter._meter_service
         self.meterId = MeterId(name, self.get_type(), tags)
 
     def get_name(self):
@@ -110,6 +112,7 @@ class BaseMeter(ABC):
         def __init__(self, name: str, tags=None):
             # Derived Builder should instantiate its corresponding meter here.
             # self.meter = BaseMeter(name, tags)
+            self.meter: Optional[BaseMeter] = None
             pass
 
         def tag(self, name: str, value):