+++ skywalking/utils/singleton.py
@@ -15,12 +15,21 @@
 # limitations under the License.
 #
 
-from skywalking.meter.gauge import Gauge
 
+class Singleton(object):
+    """
+    Written by Guido van Rossum to implement a singleton pattern.
+    https://www.python.org/download/releases/2.2/descrintro/#__new__
 
-class DataSource:
-    def register(self):
-        for name in dir(self):
-            if name.endswith('generator'):
-                generator = getattr(self, name)()
-                Gauge.Builder('instance_pvm_' + name[:-10], generator).build()
+    Classes that inherit from this class will be singletons.
+    """
+    def __new__(cls, *args, **kwds):
+        it = cls.__dict__.get('__it__')
+        if it is not None:
+            return it
+        cls.__it__ = it = object.__new__(cls)
+        it.init(*args, **kwds)
+        return it
+
+    def init(self, *args, **kwds):
+        pass