+++ tests/manual/__init__.py
@@ -1,4 +1,3 @@
-#
 # Licensed to the Apache Software Foundation (ASF) under one or more
 # contributor license agreements.  See the NOTICE file distributed with
 # this work for additional information regarding copyright ownership.
@@ -13,14 +12,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
-
-from skywalking.meter.gauge import Gauge
-
-
-class DataSource:
-    def register(self):
-        for name in dir(self):
-            if name.endswith('generator'):
-                generator = getattr(self, name)()
-                Gauge.Builder('instance_pvm_' + name[:-10], generator).build()